+++ .vscode/settings.json
@@ -1,6 +1,8 @@
 {
     "cSpell.words": [
         "openai",
-        "Rchan"
+        "Phasmophobia",
+        "Rchan",
+        "simping"
     ]
 }
