+++ src/Personality/_Personality.ts
@@ -1,11 +1,4 @@
-import { CreateChatCompletionRequest } from "openai";
-import { LolBot } from "./LolBot";
-import { Rchan } from "./Rchan";
-import { Gazelle } from "./Gazelle";
-import { Hope } from "./Hope";
-import { Joe } from "./Joe";
-import { RLol } from "./RLol";
-import { Mommy } from "./Mommy";
+import { ChatCompletionRequestMessage, ChatCompletionRequestMessageRoleEnum, CreateChatCompletionRequest } from "openai";
 import { AIDebugger } from "../AI/AIDebugger";
 
 
@@ -28,33 +21,76 @@ export enum Personalities {
     Mommy = "Mommy"
 }
 
+export class Basic implements Personality {
+    messages: Array<ChatCompletionRequestMessage> = [];
+    channel: string;
+
+    protected initialSystemMessage: string;
+    private _debug?: AIDebugger;
+
+    constructor(initialSystemMessage: string) {
+        this.initialSystemMessage = initialSystemMessage;
+        this.channel = "channel"; 
+
+        // search for messages
+        this.addSystemMessage(initialSystemMessage);
+    }
+
+    private log(str: any) {
+        if(this._debug)
+            this._debug.log(str);
+
+        else throw new Error("No debugger");
+    }
+
+    addAssistantMessage(message: string): void {
+        this.addMessage(ChatCompletionRequestMessageRoleEnum.Assistant, message);
+    }
+
+    addUserMessage(message: string, userId?: string): void {
+        this.addMessage(ChatCompletionRequestMessageRoleEnum.User, message, userId);
+    }
+
+    protected addSystemMessage(message: string) {
+        this.addMessage(ChatCompletionRequestMessageRoleEnum.System, message);
+    }
+
+    private addMessage(role: ChatCompletionRequestMessageRoleEnum, content: string, name?: string) {
+        this.messages.push({
+            role,
+            content,
+            name
+        })
+    }
+
+    getChatCompletion(): CreateChatCompletionRequest {
+        this.log(this.messages);
+        return {
+            model: "gpt-3.5-turbo",
+            messages: this.messages,
+        };
+    }
+
+    reset() {
+        this.log("Reset the personality");
+        this.messages = [];
+        this.addSystemMessage(this.initialSystemMessage);
+    }
+
+    countUserMessages() {
+        return this.messages
+            .filter( (message) => message.role == ChatCompletionRequestMessageRoleEnum.User )
+            .length
+    }
+
+    setDebugger(debug: AIDebugger) {
+        this._debug = debug;
+    }
+} 
+
 export class PersonalityFactory {
     private initBot(bot?: Personalities): Personality {
-        switch(bot) {
-            case Personalities.LOLBot:
-                return new LolBot();
-
-            case Personalities.RLol:
-                return new RLol();
-            
-            case Personalities.RChan:
-                return new Rchan();
-
-            case Personalities.Gazelle:
-                return new Gazelle();
-
-            case Personalities.Hope:
-                return new Hope();
-
-            case Personalities.Joe:
-                return new Joe();
-
-            case Personalities.Mommy:
-                return new Mommy();
-
-            default:
-                return this.initBot(Personalities.RChan);
-        }
+        
     }
 
     generateBot(debug: AIDebugger, bot?: Personalities): Personality {
